1.0.1

+++ ZTCore/source/date/ZTHandyDate.m
@@ -394,7 +394,7 @@
         case 1: case 3: case 5: case 7: case 8: case 10: case 12:
             return 31;
         case 2:
-            return [self isLeapYearWithYear:year] ? 29 : 28;
+            return [self isLeapYearWithYear:(int)year] ? 29 : 28;
     }
     
     return 30;
